renamed package to make gui tests first in list
